Create new Dropdown component
